requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
python-dotenv==1.1.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
pydantic[email]==2.11.7
httpx==0.28.1
python-multipart==0.0.20